setup.py: fixing /var/run installation when crosscompiling
When crosscompiling, python root is different from the system one, thus
the path /var/run cannot be used as is.
To obtain the path where the resource file will be created, we use
therefore the pkg_resources.resource_filename function.
This patch removes also the check for /var/run to /run symlink:
setuptools is able to create the fail2ban folder both when the
destination is a directory or a symlink

Signed-off-by: Angelo Compagnucci <angelo@amarulasolutions.com>