Kill state after adding IP to pf table
When fail2ban adds an IP to a table of the PF firewall, the state has to be killed afterwards. It this does not happen, the connection can stay open, allowing the attacker to continue attacking.
This might be unnecessary when using fail2ban to protect ssh, since the connection always resets after a failed login attempt. However, I'm using fail2ban to protect an instance of nginx against bots that cause 403 and 404 errors. During my testing, I quickly realised that without this edit, the ban would not immediately take effect.

I got the idea from this post on the FreeBSD forums:
https://forums.freebsd.org/threads/pf-and-fail2ban.51624/