Fix cloudflare-token.conf
fix on nixos bug:
uses unban error:
 `id for ip cannot be found using target`

The reason why the error occurs on nixos is that the special operating environment of nixos causes the commands before and after it to be executed separately. Subsequent commands do not inherit the previous command. Merging it solves the problem without affecting other distributions.