Fix expansion of pre-rule parameter
If the `pre-rule` paramater is used -- i.e. not empty -- it is expanded without a separating whithespace to the next parmeters -- typical `-p ...`. This causes the call of the `iptables` command to fail.

Just add a space to separate the expanded variables; the superfluous space in case of an empty pre-role seems tolerable to me.