make Ticket.getID() return str
If the original id is a str, it's internally stored as IPAddr. Instead
of returning the IPAddr, the getter now only returns a plain str. If it
originally was a tuple or None, that is passed as-is. Also now throws an
error if the id has an unexpected type. One testcase now fails.